refactor: move redis initialization to a separate shared util
